refactor: BlockSerializable to RpcBlock
