+++ crates/rpc/types/mod.rs
@@ -1,1 +1,2 @@
+pub mod block;
 pub mod transaction;
